+++ buildSrc/src/main/groovy/com/worksap/nlp/tools/engines.groovy
@@ -11,7 +11,8 @@ enum EsSupport implements EngineSupport {
     Es80("es-8.00"),
     Es83("es-8.03"),
     Es84("es-8.04"),
-    Es810("es-8.10")
+    Es810("es-8.10"),
+    Es812("es-8.12"),
 
     String tag
     List<String> keys
@@ -36,8 +37,10 @@ enum EsSupport implements EngineSupport {
             return Es83
         } else if (vers.ge(8, 4) && vers.lt(8, 10)) {
             return Es84
-        } else if (vers.ge(8, 10) && vers.lt(9, 0)) {
+        } else if (vers.ge(8, 10) && vers.lt(8, 12)) {
             return Es810
+        } else if (vers.ge(8, 12) && vers.lt(9, 0)) {
+            return Es812
         } else {
             throw new IllegalArgumentException("unsupported ElasticSearch version: " + vers.raw)
         }
