+++ integration/src/test/ext/es-8.00-ge/com.worksap.nlp.elasticsearch.sudachi/aliases.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Works Applications Co., Ltd.
+ * Copyright (c) 2023-2024 Works Applications Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,37 +16,18 @@
 
 package com.worksap.nlp.elasticsearch.sudachi
 
-import org.elasticsearch.action.admin.indices.analyze.AnalyzeAction
-import org.elasticsearch.action.admin.indices.analyze.TransportAnalyzeAction
-import org.elasticsearch.common.settings.Settings
-import org.elasticsearch.env.Environment
-import org.elasticsearch.indices.analysis.AnalysisModule
-import org.elasticsearch.plugins.AnalysisPlugin
-import org.elasticsearch.plugins.PluginsService
-
 typealias SearchEngineTestCase = org.elasticsearch.test.ESTestCase
-
-typealias PluginsServiceAlias = PluginsService
 
 typealias AnalysisRegistryAlias = org.elasticsearch.index.analysis.AnalysisRegistry
 
-typealias EnvironmentAlias = Environment
+typealias EnvironmentAlias = org.elasticsearch.env.Environment
 
-typealias SettingsAlias = Settings
+typealias SettingsAlias = org.elasticsearch.common.settings.Settings
 
-typealias AnalysisPluginAlias = AnalysisPlugin
+typealias AnalysisPluginAlias = org.elasticsearch.plugins.AnalysisPlugin
 
-typealias AnalyzeActionRequestAlias = AnalyzeAction.Request
+typealias AnalyzeActionRequestAlias =
+    org.elasticsearch.action.admin.indices.analyze.AnalyzeAction.Request
 
-typealias TransportAnalyzeActionAlias = TransportAnalyzeAction
-
-fun SudachiInSearchEngineEnv.makePluginService(): PluginsServiceAlias {
-  return PluginsService(settings(), configPath, null, pluginsPath)
-}
-
-fun SudachiInSearchEngineEnv.makeAnalysisModule(): AnalysisModule {
-  val plugins = makePluginService()
-  val analysisPlugins = plugins.filterPlugins(AnalysisPlugin::class.java)
-  val env = environment()
-  return AnalysisModule(env, analysisPlugins, plugins.stablePluginRegistry)
-}
+typealias TransportAnalyzeActionAlias =
+    org.elasticsearch.action.admin.indices.analyze.TransportAnalyzeAction
