+++ integration/src/test/java/com/worksap/nlp/elasticsearch/sudachi/BasicTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Works Applications Co., Ltd.
+ * Copyright (c) 2023-2024 Works Applications Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@ class BasicTest : SudachiEnvTest() {
   @Test
   fun canLoadPlugin() {
     val plugins = sudachiEnv.makePluginService()
-    val analysisPlugins = plugins.filterPlugins(AnalysisPluginAlias::class.java)
+    val analysisPlugins = plugins.filterPlugins(AnalysisPluginAlias::class.java).toList()
     Assert.assertEquals(2, analysisPlugins.size)
     val plugin =
         analysisPlugins.find {
